+++ SomnusBackend/src/main/java/com/somnus/server/backend/users/UserController.java
@@ -2,10 +2,7 @@ package com.somnus.server.backend.users;
 
 import com.somnus.server.backend.users.dto.UserDto;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping(value = "/user-api")
@@ -17,6 +14,11 @@ public class UserController {
     @GetMapping(value = "/authentication/authenticate-user")
     public UserDto authenticateUser(@RequestHeader String firebaseToken){
         return userService.authenticateUser(firebaseToken);
+    }
+
+    @PostMapping(value = "/authentication/register-user")
+    public UserDto registerUser(@RequestHeader String firebaseToken, @RequestBody UserDto userDto){
+        return userService.registerUser(firebaseToken, userDto);
     }
 
 }